+++ ext/impex.ext.console.js
@@ -1,0 +1,69 @@
+/**
+ * impex日志系统
+ */
+;!function(g){
+
+    g.imLog = new function(){
+        this.level = 1;
+        this.warn = function(txt){
+            if(this.level < 2)return;
+            console.warn('im warn :: ' + txt);
+        }
+        this.error = function(txt){
+            if(this.level < 1)return;
+            console.error('im error :: ' + txt);
+        }
+        this.debug = function(txt){
+            if(this.level < 3)return;
+            console.debug('im debug :: ' + txt);
+        }
+        this.log = function(txt){
+            if(this.level < 4)return;
+
+            if(arguments.length===2){
+                stateLog.call(null,arguments[0],arguments[1]);
+                return;
+            }
+            console.log('im log :: ' + txt);
+        }
+    }
+
+    var lastComp;
+    function stateLog(comp,state){
+        var indent = '';
+        var p = comp.$parent;
+        while(p){
+            indent += '=';
+            p = p.$parent;
+        }
+        var info = '';
+        if(comp === lastComp){
+            info = '↑↑↑ ↑↑↑ ↑↑↑ ';
+        }else{
+            var props = [];
+            props.push('id:'+comp.$__id);
+            if(comp.$name)
+                props.push('name:'+comp.$name);
+            var viewName = comp.$view?comp.$view.elements[0].tagName:'';
+            props.push('view:'+viewName);
+            if(comp.$parent){
+                props.push('parentId:'+(comp.$parent?comp.$parent.$__id:'null'));
+            }
+
+            var type = comp.$endTag === null?'Directive':'Component';
+
+            info = type+'{'+ props.join(',') +'} ';
+        }
+        lastComp = comp;
+        
+        console.log('im log :: '+indent + (indent?'> ':'') + info + state);
+    }
+
+    //polyfill
+    self.console = self.console||new function(){
+        this.log = function(){}
+        this.debug = function(){}
+        this.error = function(){}
+        this.warn = function(){}
+    }
+}(window);