+++ src/factory/Factory.js
@@ -15,11 +15,13 @@ Factory.prototype = {
 		type = type.toLowerCase();
 
 		//keywords check
-		var ks = Object.keys(model);
-		for(var i=BUILD_IN_PROPS.length;i--;){
-			if(ks.indexOf(BUILD_IN_PROPS[i])>-1){
-				impex.console.error('attempt to overwrite build-in property['+BUILD_IN_PROPS[i]+'] of Component['+type+']');
-				return;
+		if(this.baseClass === Component || this.baseClass === Directive ){
+			var ks = Object.keys(model);
+			for(var i=BUILD_IN_PROPS.length;i--;){
+				if(ks.indexOf(BUILD_IN_PROPS[i])>-1){
+					LOGGER.error('attempt to overwrite build-in property['+BUILD_IN_PROPS[i]+'] of Component['+type+']');
+					return;
+				}
 			}
 		}